+++ src/controllers/project.js
@@ -2,13 +2,21 @@
 
 const { Project, validateProject } = require("../models/project");
 const mongoose = require("mongoose");
-const { Selection } = require("../models/selection");
-
 /**
  * Get all projects
  * @returns {array} list of all projects
  */
 const getAll = async () => await Project.find().sort("name");
+
+/**
+ * Get all projects by list id
+ * @returns {array} list of all projects
+ */
+const getAllByListId = async (list_id) => {
+  let objectIds = [];
+  list_id.map((id) => objectIds.push(mongoose.Types.ObjectId(id)));
+  return await Project.find({ _id: { $in: objectIds } }).sort("name");
+};
 
 /**
  * Get project by id
@@ -18,13 +26,21 @@ const getAll = async () => await Project.find().sort("name");
 const getById = async (id) =>
   await Project.findById(mongoose.Types.ObjectId(id));
 
-const addSelection = async (selection, id) => {
+/**
+ * Add selection to your respective project
+ * @param {object} selection
+ */
+const addSelection = async (selection) => {
   let project = await getById(selection.projectId);
   project = { ...project._doc };
   project.selections.push(selection._id);
   await update(project._id, project);
 };
 
+/**
+ * Remove selection to your respective project
+ * @param {string} selectionId
+ */
 const removeSelection = async (selectionId) => {
   let project = await Project.findOne({ selections: selectionId.toString() });
   project = { ...project._doc };
@@ -76,8 +92,21 @@ const update = async (id, { name, description, labId, selections }) =>
  * @param {string} id
  * @returns {object} project removed
  */
-const remove = async (id) =>
-  await Project.findByIdAndRemove(mongoose.Types.ObjectId(id));
+const remove = async (id) => {
+  const project = await Project.findById(mongoose.Types.ObjectId(id));
+  const SelectionController = require("./selection");
+
+  if (project) {
+    await Promise.all(
+      project.selections.map(
+        async (selectionId) => await SelectionController.remove(selectionId)
+      )
+    );
+    return await Project.findByIdAndRemove(mongoose.Types.ObjectId(id));
+  } else {
+    throw new Error("The project with the given ID was not found.");
+  }
+};
 
 /**
  * Validate project
@@ -91,6 +120,7 @@ const validate = (object) => {
 
 module.exports = {
   getAll,
+  getAllByListId,
   getById,
   create,
   update,
